Tracking changes

+++ aspnet-core/src/TACHYON.Application.Shared/Tracking/ITrackingAppService.cs
@@ -7,13 +7,13 @@ using TACHYON.Tracking.Dto;
 
 namespace TACHYON.Tracking
 {
-    public interface ITrackingAppService:IApplicationService
+    public interface ITrackingAppService : IApplicationService
     {
         Task<PagedResultDto<TrackingListDto>> GetAll(TrackingSearchInputDto Input);
         Task<ListResultDto<ShippingRequestTripDriverRoutePointDto>> GetForView(long id);
         Task Accept(int id);
         Task Start(int id);
-        Task ChangeStatus(long id);
+        Task ChangeStatus(int id);
         Task NextLocation(long id);
         Task ConfirmReceiverCode(ConfirmReceiverCodeInput input);
         Task<FileDto> POD(long id);
